Simple Dijkstra algorithm
This is not a fully functional piece of code. In order to function, a network of nodes needs to be identified, and the node structure needs to be declared. In testing, I was manually filling everything in before the main function runs. So, no demonstration of this is currently possible without manually filling in those values.

Not only that, but I think I know a better way of doing this that can handle directions better, but I didn't have time to get it working before our demo Monday. So expect to see this file drastically change very soon.